requirements.txt
# Основные зависимости для RAG решения
pandas==2.3.3
numpy==2.3.4
scikit-learn==1.7.2

# API клиенты
openai==2.6.1

# Утилиты
python-dotenv==1.2.1
tqdm==4.67.1

# Логирование и отладка (опционально)
matplotlib==3.9.3
seaborn==0.13.2

# Для продвинутых функций (опционально)
# sentence-transformers==5.1.2  # если захотите использовать локальные модели
# faiss-cpu==1.12.0  # для более быстрого поиска в больших базах
